restructure the module path
