+++ aiidalab_qe/__init__.py
@@ -1,11 +1,11 @@
 """Package for the AiiDAlab QE app."""
 
-from aiidalab_qe.process import WorkChainSelector
-from aiidalab_qe.steps import (
+from aiidalab_qe.app.process import WorkChainSelector
+from aiidalab_qe.app.steps import (
     SubmitQeAppWorkChainStep,
     ViewQeAppWorkChainStatusAndResultsStep,
 )
-from aiidalab_qe.structures import StructureSelectionStep
+from aiidalab_qe.app.structures import StructureSelectionStep
 from aiidalab_qe.version import __version__
 
 __all__ = [
