+++ aiidalab_qe/__main__.py
@@ -1,8 +1,11 @@
+"""For running the app from the command line used for post_install script.
+"""
+
 import click
 
-from .setup_codes import codes_are_setup
-from .setup_codes import install as install_qe_codes
-from .sssp import install as setup_sssp
+from aiidalab_qe.app.setup_codes import codes_are_setup
+from aiidalab_qe.app.setup_codes import install as install_qe_codes
+from aiidalab_qe.app.sssp import install as setup_sssp
 
 
 @click.group()
