+++ aiidalab_qe/app/parameters/__init__.py
@@ -2,6 +2,6 @@ from importlib import resources
 
 import yaml
 
-from aiidalab_qe import parameters
+from aiidalab_qe.app import parameters
 
 DEFAULT_PARAMETERS = yaml.safe_load(resources.read_text(parameters, "qeapp.yaml"))
