+++ aiidalab_qe/app/steps.py
@@ -24,11 +24,11 @@ from aiidalab_widgets_base import (
 )
 from IPython.display import display
 
-from aiidalab_qe.parameters import DEFAULT_PARAMETERS
-from aiidalab_qe.pseudos import PseudoFamilySelector
-from aiidalab_qe.setup_codes import QESetupWidget
-from aiidalab_qe.sssp import SSSPInstallWidget
-from aiidalab_qe.widgets import ParallelizationSettings, ResourceSelectionWidget
+from aiidalab_qe.app.parameters import DEFAULT_PARAMETERS
+from aiidalab_qe.app.pseudos import PseudoFamilySelector
+from aiidalab_qe.app.setup_codes import QESetupWidget
+from aiidalab_qe.app.sssp import SSSPInstallWidget
+from aiidalab_qe.app.widgets import ParallelizationSettings, ResourceSelectionWidget
 from aiidalab_qe.workflows import QeAppWorkChain
 
 StructureData = DataFactory("core.structure")
