+++ aiidalab_qe/version.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-"""This module contains project version information."""
+"""This module contains project version information for both the app and the workflow."""
 
 __version__ = "v23.04.1"
